Split attendee-answers into save.ts + reads.ts (under 400 lines)
Continues PR #1678's questions.ts split. attendee-answers.ts had grown to
481 lines, over AGENTS.md's ~400 target, so it's split into a folder:

- attendee-answers/save.ts  — saveAttendeeAnswers, groupListingAnswers,
  AttendeeAnswerSet, AttendeeListingEntry + save-only helpers (245 lines)
- attendee-answers/reads.ts  — batch + single reads (244 lines)

All callers migrated to the new paths (no compat re-exports). Deleted the
old monolith. Verified: typecheck, strict lint, jscpd 0%, and the affected
test files incl. the dedicated attendee-answers suite and server-attendees.

Addresses Codex P2 review thread on attendee-answers.ts.
