+++ src/features/admin/group-page-data.ts
@@ -22,7 +22,7 @@ import {
 import { getActiveHolidays } from "#shared/db/holidays.ts";
 import { getGroupDayPrices } from "#shared/db/listing-prices.ts";
 import { getAttendeesByListingIds } from "#shared/db/listings.ts";
-import { loadAttendeeQuestionData } from "#shared/db/questions/attendee-answers.ts";
+import { loadAttendeeQuestionData } from "#shared/db/questions/attendee-answers/reads.ts";
 import { settings } from "#shared/db/settings.ts";
 import { requireRequestPrivateKey } from "#shared/session-private-key.ts";
 import { sortListings } from "#shared/sort-listings.ts";
