+++ src/shared/merge/attendee-merge.ts
@@ -22,8 +22,8 @@ import type { QuestionWithAnswers } from "#shared/db/question-types.ts";
 import {
   getAttendeeAnswersByQuestion,
   getAttendeeTextAnswers,
-  saveAttendeeAnswers,
-} from "#shared/db/questions/attendee-answers.ts";
+} from "#shared/db/questions/attendee-answers/reads.ts";
+import { saveAttendeeAnswers } from "#shared/db/questions/attendee-answers/save.ts";
 import type { AccountRef } from "#shared/ledger/types.ts";
 import type {
   ApplyAttendeeMergeInput,
