+++ src/ui/templates/attendee-table.tsx
@@ -25,7 +25,7 @@ import {
 } from "#shared/columns/attendee-columns.ts";
 import { isServicing } from "#shared/db/attendees/kind.ts";
 import type { QuestionWithAnswers } from "#shared/db/question-types.ts";
-import type { AttendeeQuestionData } from "#shared/db/questions/attendee-answers.ts";
+import type { AttendeeQuestionData } from "#shared/db/questions/attendee-answers/reads.ts";
 import { settings } from "#shared/db/settings.ts";
 import { CsrfForm } from "#shared/forms.tsx";
 import {
