+++ test/lib/server-parents-gate.test.ts
@@ -2,7 +2,7 @@ import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
 import { getAttendeesRaw } from "#shared/db/attendees.ts";
 import { listingChildren } from "#shared/db/listing-parents.ts";
-import { getAttendeeAnswersBatch } from "#shared/db/questions/attendee-answers.ts";
+import { getAttendeeAnswersBatch } from "#shared/db/questions/attendee-answers/reads.ts";
 import { setListingQuestions } from "#shared/db/questions/queries.ts";
 import { answersTable, questionsTable } from "#shared/db/questions/tables.ts";
 import {
