+++ test/shared/db/questions/helpers.ts
@@ -1,6 +1,6 @@
 import { expect } from "@std/expect";
 import type { Question, TextAnswer } from "#shared/db/question-types.ts";
-import { saveAttendeeAnswers } from "#shared/db/questions/attendee-answers.ts";
+import { saveAttendeeAnswers } from "#shared/db/questions/attendee-answers/save.ts";
 import { setListingQuestions } from "#shared/db/questions/queries.ts";
 import { assignNextQuestionSortOrder } from "#shared/db/questions/sort-order.ts";
 import { answersTable, questionsTable } from "#shared/db/questions/tables.ts";
