+++ test/shared/merge/attendee-merge/helpers.ts
@@ -7,7 +7,7 @@ import { postTransfers } from "#shared/accounting/store.ts";
 import { LISTING_ATTENDEE_ROW_COLS } from "#shared/db/attendees.ts";
 import { queryAll } from "#shared/db/client.ts";
 import type { QuestionWithAnswers } from "#shared/db/question-types.ts";
-import { saveAttendeeAnswers } from "#shared/db/questions/attendee-answers.ts";
+import { saveAttendeeAnswers } from "#shared/db/questions/attendee-answers/save.ts";
 import { setListingQuestions } from "#shared/db/questions/queries.ts";
 import { answersTable, questionsTable } from "#shared/db/questions/tables.ts";
 import {
